requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
pydantic==1.10.8
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiosqlite==0.19.0
pytest==7.4.3
httpx==0.25.2
Jinja2==3.1.2
python-dotenv==1.0.0

# LLM and NLP dependencies
openai>=1.0.0
